# Conflicts:
#	makesymlinks.sh

+++ makesymlinks.sh
@@ -30,7 +30,7 @@ for file in $files; do
     ln -s $dir/$file ~/.$file
 done
 
-function install_zsh {
+install_zsh () {
 # Test to see if zshell is installed.  If it is:
 if [ -f /bin/zsh -o -f /usr/bin/zsh ]; then
     # Clone my oh-my-zsh repository from GitHub only if it isn't already present
@@ -46,8 +46,14 @@ else
     platform=$(uname);
     # If the platform is Linux, try an apt-get to install zsh and then recurse
     if [[ $platform == 'Linux' ]]; then
-        sudo apt-get install zsh
-        install_zsh
+        if [[ -f /etc/redhat-release ]]; then
+            sudo yum install zsh
+            install_zsh
+        fi
+        if [[ -f /etc/debian_version ]]; then
+            sudo apt-get install zsh
+            install_zsh
+        fi
     # If the platform is OS X, tell the user to install zsh :)
     elif [[ $platform == 'Darwin' ]]; then
         echo "Please install zsh, then re-run this script!"
